app/requirements.txt
asgiref==3.6.0
beautifulsoup4==4.12.0
boto3==1.26.104
botocore==1.29.104
cffi==1.15.1
cryptography==40.0.1
Django==4.0.7
django-admin-interface==0.26.1
django-colorfield==0.11.0
django-cors-headers==3.14.0
django-extensions==3.2.3
django-filter==23.1
django-rename-app==0.1.6
django-rest-knox==4.2.0
django-storages==1.11.1
djangorestframework==3.14.0
gunicorn==20.0.0
Jinja2==3.1.3
jmespath==1.0.1
Markdown==3.3.6
MarkupSafe==2.1.5
mock==3.0.5
Pillow==9.3.0
psycopg2-binary==2.8.5
pycparser==2.21
python-dateutil==2.8.2
python-slugify==8.0.3
pytz==2019.3
s3transfer==0.6.0
six==1.13.0
soupsieve==2.4
sqlparse==0.3.0
text-unidecode==1.3
urllib3==1.26.15
imgkit
